♻️ refactor: Refatora módulo Intelesign
- Atualiza dependências do projeto
- Refatora DTOs e entidades para melhor organização
- Remove arquivos desnecessários
- Simplifica a lógica do serviço Intelesign
