+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "alertanow",
     "AMERICO",
     "amqp",
+    "Annot",
     "apiv",
     "aprovacao",
     "argb",
@@ -81,6 +82,7 @@
     "Messager",
     "meuarquivo",
     "Millis",
+    "minetype",
     "minioadmin",
     "nestjs",
     "nomearquivo",
@@ -119,6 +121,7 @@
     "razaosocial",
     "rbrp",
     "redebrasilrp",
+    "Reistro",
     "relat",
     "relatorio",
     "relatoriofinanceiro",
@@ -133,6 +136,7 @@
     "SERVICOS",
     "signatario",
     "signatarios",
+    "siguinatario",
     "sisapp",
     "sisnato",
     "sisnatodoc",
