+++ package.json
@@ -46,6 +46,7 @@
     "@pdf-lib/fontkit": "^1.1.1",
     "@prisma/client": "^6.16.1",
     "@prisma/instrumentation": "^6.16.1",
+    "@types/form-data": "^2.2.1",
     "@types/passport-jwt": "^4.0.1",
     "amqp-connection-manager": "^4.1.14",
     "amqplib": "^0.10.8",
@@ -55,6 +56,7 @@
     "class-validator": "^0.14.1",
     "dotenv": "^17.2.1",
     "exceljs": "^4.4.0",
+    "form-data": "^4.0.4",
     "mime-types": "^2.1.35",
     "mysql2": "^3.14.0",
     "passport": "^0.7.0",
