+++ src/api/intelesign/dto/create-intelesign.dto.ts
@@ -1,23 +1,18 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
-import {
-  IsArray,
-  IsEnum,
-  IsNumber,
-  IsOptional,
-  IsString,
-} from 'class-validator';
+import { IsNumber, IsOptional } from 'class-validator';
+import { SignatarioDto } from './sign.dto';
 
 export class CreateIntelesignDto {
   @ApiProperty({
     description: 'Array de IDs dos signatários',
     example: '[]',
     required: false,
-    type: () => [String],
+    type: () => [SignatarioDto],
   })
   @IsOptional()
   @Transform(({ value }) => JSON.parse(value))
-  signatarios: SignatarioDto[];
+  signatarios?: SignatarioDto[];
 
   @ApiProperty({
     description: 'Valor do documento',
@@ -49,52 +44,4 @@ export class CreateIntelesignDto {
     },
   )
   cca_id?: number;
-}
-
-
-export class SignatarioDto {
-  @ApiProperty({
-    description: 'ID do signatário',
-    example: '1',
-    required: true,
-    type: () => Number,
-  })
-  @Transform(({ value }) => Number(value))
-  @IsNumber(
-    {},
-    {
-      message: 'ID deve ser um número',
-    },
-  )
-  id: number;
-
-  @ApiProperty({
-    description: 'Tipo de assinatura',
-    example: 'simple || qualified',
-    required: true,
-    enum: ['simple', 'qualified'],
-    type: () => String,
-  })
-  @IsString({
-    message: 'AssType deve ser uma string',
-  })
-  @IsEnum(['simple', 'qualified'], {
-    message: 'AssType deve ser simple ou qualified',
-  })
-  asstype: string;
-
-  @ApiProperty({
-    description: 'Tipo de destinatário',
-    example: 'signer || approver || carbon-copy',
-    required: true,
-    enum: ['signer', 'approver', 'carbon-copy'],
-    type: () => String,
-  })
-  @IsEnum(['signer', 'approver', 'carbon-copy'], {
-    message: 'Type deve ser signer, approver ou carbon-copy',
-  })
-  @IsString({
-    message: 'Type deve ser uma string',
-  })
-  type: string;
 }
